cleaned up resnet code
