+++ agent/lib/observation_normalizer.py
@@ -1,0 +1,50 @@
+import torch
+from torch import nn
+
+
+# ##ObservationNormalization
+# These are approximate maximum values for each feature. Ideally they would be defined closer to their source,
+# but here we are. If you add / remove a feature, you should add / remove the corresponding normalization.
+OBS_NORMALIZATIONS = {
+    'agent': 1,
+    'agent:hp': 1,
+    'agent:frozen': 1,
+    'agent:energy': 255,
+    'agent:orientation': 1,
+    'agent:shield': 1,
+    'agent:inv:r1': 5,
+    'agent:inv:r2': 5,
+    'agent:inv:r3': 5,
+    'wall': 1,
+    'wall:hp': 10,
+    'generator': 1,
+    'generator:hp': 30,
+    'generator:r1': 30,
+    'generator:ready': 1,
+    'converter': 1,
+    'converter:hp': 30,
+    'converter:input_resource': 5,
+    'converter:output_resource': 5,
+    'converter:output_energy': 100,
+    'converter:ready': 1,
+    'altar': 1,
+    'altar:hp': 30,
+    'altar:ready': 1,
+    'last_action': 10,
+    'last_action_argument': 10,
+    'agent:kinship': 10,
+}
+
+class ObservationNormalizer(nn.Module):
+    def __init__(self, grid_features: list[str]):
+        super().__init__()
+
+        num_objects = len(grid_features)
+        
+        obs_norm = torch.tensor([OBS_NORMALIZATIONS[k] for k in grid_features], dtype=torch.float32)
+        obs_norm = obs_norm.view(1, num_objects, 1, 1)
+
+        self.register_buffer('obs_norm', obs_norm)
+
+    def forward(self, obs):
+        return obs / self.obs_norm
