+++ configs/user/gabrielk.yaml
@@ -1,0 +1,44 @@
+# @package __global__
+
+seed: null
+
+defaults:
+  - override /hardware: aws
+  - override /env/mettagrid@env: simple
+  - override /agent: resnet
+
+# policy: file://.train_dir/gabrielk.local.train.simple/checkpoints/model_0000.pt
+# baselines: file://.train_dir/gabrielk.local.train.simple/checkpoints/model_0000.pt
+
+# evaluator:
+#   policy:
+#     uri: ${...policy_uri}
+  # baselines:
+  #   uri: ${...baseline_uri}
+
+policy_uri: file://./train_dir/gabrielk.local.train.resnet/checkpoints/model_0000.pt
+baseline_uri: file://./train_dir/gabrielk.local.train.simple/checkpoints/model_0000.pt
+
+env:
+  game:
+    max_steps: 100
+
+trainer:
+  evaluate_interval: 2
+
+evaluator:
+  policy:
+    uri: ${...policy_uri}
+  baselines:
+    uri: ${...baseline_uri}
+
+wandb:
+  enabled: false
+  track: false
+  checkpoint_interval: 1
+
+sweep:
+  metric: action.use
+
+cmd: ???
+run: ${oc.env:USER}.local.${cmd}.simple
